moved module creation logic, including Sample.module, from cli to cli-lib
