+++ packages/cli-lib/defaults/Sample.module/fields.json
