+++ packages/cli-lib/defaults/Sample.module/meta.json
