+++ packages/cli-lib/defaults/Sample.module/module.html
